requirements.txt
asgiref==3.7.1
Django==4.2
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
pillow==10.2.0
PyJWT==2.8.0
sqlparse==0.4.4
typing_extensions==4.6.2
